+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 .. note:: This version is not yet released and is under active development.
 
 * Drop support for python 2.6 and 3.2
+* Define minimal pyparsing version to 2.0.2 (:issue:`91`).
 
 
 16.8 - 2016-10-29