+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 ---------
 
+17.1 - `master`_
+~~~~~~~~~~~~~~~~
+
+.. note:: This version is not yet released and is under active development.
+
+
 17.0 - 2017-02-28
 ~~~~~~~~~~~~~~~~~
 
@@ -179,3 +185,6 @@ Changelog
 ~~~~~~~~~~~~~~~~~
 
 * Initial release.
+
+
+.. _`master`: https://github.com/pypa/packaging/